Add files via upload
plantUML code for the Object Diagrams. They should represent the situation in which "HelloWorldCS4800" was just added, without any changes being made to the CharacterProperties, and a different situation after the Characters H and W received changes to their CharacterProperties.

NOTE: there are two different versions for the solutions, one set in which every single character is represented as an object, and the other with only H and W being represented as objects because they are the only characters that have their properties changed. This was done because the version with every single character being represented as an object appears very crowded and can be hard to read.

before_modifications.puml and after_modified_H_W.puml are the object diagrams with every single Character within "HelloWorldCS4800" represented as an object.

editor_before.puml and editor_after.puml are the object diagrams where only H and W are represented as Character objects for the sake of space and simplicity because they are the only characters that are receiving changes to their CharacterProperties